requirements.txt
# Core dependencies for trading bot project

# Data collection
yfinance>=0.2.18
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
lxml>=4.9.0

# Environment management
python-dotenv>=1.0.0

# Machine Learning and NLP
openai>=1.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Backtesting
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pathlib2>=2.3.7

# Optional: Enhanced data sources
# alpha-vantage>=2.3.1
# quandl>=3.7.0
# sec-api>=1.0.0

# Development tools (optional)
# jupyter>=1.0.0
# ipykernel>=6.20.0